Fixed connection to mongo, was causing an error in file upload
Fixed connection to mongo, was causing an error in file upload
