+++ conf/deployment_configs.json
@@ -7,7 +7,7 @@
     "eudatToken" : "__FIXME",
     "sendGridUser" : "__FIXME",
     "sendGridPassword" : "__FIXME",
-    "elasticSearchHost" : "192.168.56.101",
+    "elasticSearchHost" : "192.168.56.249",
     "elasticSearchPort" : "9200",
     "cache": {
       "active": true,
@@ -15,7 +15,7 @@
       {
         "options":
         {
-          "host" : "192.168.56.101",
+          "host" : "192.168.56.249",
           "port" : "6379"
         },
         "database_number" : 1
@@ -26,13 +26,13 @@
         "cache_period_in_seconds" : 3600
       }
     },
-    "virtuosoHost" : "192.168.56.101",
+    "virtuosoHost" : "192.168.56.249",
     "virtuosoPort" : "8890",
     "virtuosoAuth" : {
       "user" : "dba",
       "password" : "root"
     },
-    "mongoDBHost" : "192.168.56.101",
+    "mongoDBHost" : "192.168.56.249",
     "mongoDbPort" : "27017",
     "mongoDbCollectionName" : "dendro_data_dev",
     "mongoDbVersion" : "",
@@ -40,13 +40,13 @@
       "user" : "",
       "password" : ""
     },
-    "mySQLHost" : "192.168.56.101",
+    "mySQLHost" : "192.168.56.249",
     "mySQLPort" : "3306",
     "mySQLAuth" : {
       "user" : "root",
-      "password" : "root"
+      "password" : "r00t_p4ssw0rd"
     },
-    "mySQLDBName" : "dev",
+    "mySQLDBName" : "dendroVagrantDemo",
     "maxUploadSize" : 200000000,
     "maxProjectSize" : 200000000,
     "maxSimultanousConnectionsToDb" : 3,
